refactor: move veil to separate file
