+++ js/ui/export-popup.js
@@ -1,4 +1,4 @@
-import { veil } from '../uicode.js'
+import { veil } from './veil.js'
 import { PopupHeader } from './popup-header.js'
 import { globalPopup } from '../app.js'
 import { Store } from '../store.js'
